Minor changes: renamed a misnamed method and changed an int constant to double to avoid having to cast it to a double everywhere.

+++ Glyssen/Character/CharacterGroupList.cs
@@ -122,7 +122,7 @@ namespace Glyssen.Character
 					if (keyStrokesByCharacterId.TryGetValue(characterId, out keystrokesForCharacter))
 						keyStrokes += keystrokesForCharacter;
 				}
-				group.EstimatedHours = keyStrokes / (double)Program.kKeyStrokesPerHour;
+				group.EstimatedHours = keyStrokes / Program.kKeyStrokesPerHour;
 			}
 		}
 	}
